tidy predicate encoder
